refactor client, common, server
