--- webpack.config.dev.js
@@ -1,63 +1,0 @@
-// jscs:disable
-var path = require('path')
-var webpack = require('webpack')
-var AssetsPlugin = require('assets-webpack-plugin')
-
-module.exports = {
-  devtool: 'source-map',
-  entry: {
-    main: [
-      'webpack-hot-middleware/client',
-      './src/client.js'
-    ],
-    vendor: [
-      'react',
-      'react-dom',
-      'react-router',
-      'redux',
-      'react-redux',
-      'aphrodite'
-    ]
-  },
-  output: {
-    path: path.join(__dirname, 'dist'),
-    filename: '[name].js',
-    chunkFilename: '[id].chunk.js',
-    publicPath: '/build/static/'
-  },
-  plugins: [
-    new webpack.optimize.OccurrenceOrderPlugin(),
-    new webpack.HotModuleReplacementPlugin(),
-    new webpack.optimize.CommonsChunkPlugin('vendor', 'vendor.js', 2),
-    new webpack.NoErrorsPlugin(),
-    new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify('development'),
-      '__DEV__': true
-    })
-  ],
-  module: {
-    preLoaders: [
-      {
-        // set up standard-loader as a preloader
-        test: /\.js?$/,
-        loader: 'standard',
-        include: path.join(__dirname, 'src')
-      }
-    ],
-    loaders: [
-      {
-        test: /\.js$/,
-        loader: 'babel',
-        query: {
-          cacheDirectory: true,
-          presets: ["es2015", "react", "stage-0"]
-        },
-        include: path.join(__dirname, 'src')
-      }
-    ]
-  },
-  standard: {
-    // config options to be passed through to standard e.g.
-    parser: 'babel-eslint'
-  }
-}
